Add working random level generation
This commit adds the working level generation.
