+++ src/main/java/nl/tudelft/contextproject/model/level/RandomLevelFactory.java
@@ -1,12 +1,20 @@
 package nl.tudelft.contextproject.model.level;
 
+import java.io.File;
+import java.net.URISyntaxException;
 import java.util.ArrayList;
+import java.util.List;
 import java.util.Random;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import com.jme3.light.Light;
 import com.jme3.light.PointLight;
 import com.jme3.math.ColorRGBA;
+import com.jme3.math.Vector2f;
 import com.jme3.math.Vector3f;
+import lombok.SneakyThrows;
+import nl.tudelft.contextproject.util.Size;
 
 
 /**
@@ -14,30 +22,42 @@ import com.jme3.math.Vector3f;
  * NOTE: This factory becomes deprecated after a proper factory is introduced!
  */
 public class RandomLevelFactory implements LevelFactory {
+	private static final int MAX_WIDTH = 50;
+	private static final int MAX_HEIGHT = 50;
+	private static final int MAX_ATTEMPTS = 10;
+	private static final Pattern PATTERN = Pattern.compile("(?<width>\\d+)x(?<height>\\d+)_.*");
+
 	private Random rand;
-	private int width;
-	private int height;
+	private int amount;
+	private boolean allowDuplicates;
 
 	/**
 	 * Constructor for the random level factory.
-	 * @param width The width of the maze.
-	 * @param height The height of the maze.
+	 *
+	 * @param amount
+	 * 			the amount of rooms to generate
+	 * @param allowDuplicates
+	 * 			if duplicates are allowed
 	 */
-	public RandomLevelFactory(int width, int height) {
-		this.height = height;
-		this.width = width;
+	public RandomLevelFactory(int amount, boolean allowDuplicates) {
+		this.amount = amount;
+		this.allowDuplicates = allowDuplicates;
 	}
 
 	@Override
 	public Level generateSeeded(long seed) {
 		rand = new Random(seed);
 
-		MazeTile[][] mazeTiles = new MazeTile[width][height];
+		List<GeneratorRoom> generated = create();
+		TileType[][] carved = carveRooms(generated);
+		carved = carveCorridors(carved, generated);
 
-		for (int x = 0; x < width; x++) {
-			for (int y = 0; y < height; y++) {
-				if (rand.nextFloat() < .3f) {
-					mazeTiles[x][y] = new MazeTile(x, y);
+		MazeTile[][] mazeTiles = new MazeTile[MAX_WIDTH][MAX_HEIGHT];
+
+		for (int x = 0; x < carved.length; x++) {
+			for (int y = 0; y < carved[0].length; y++) {
+				if (carved[x][y] != null) {
+					mazeTiles[x][y] = new MazeTile(x, y, carved[x][y]);
 				}
 			}
 		}
@@ -49,6 +69,216 @@ public class RandomLevelFactory implements LevelFactory {
 		p.setRadius(20);
 		lights.add(p);
 		return new Level(mazeTiles, lights);
+	}
+
+	/**
+	 * Return a random number between two values.
+	 * @param min
+	 *          The minimum value the random number can be.
+	 * @param max
+	 *          The maximum value the random number can be.
+	 * @return
+	 *          The random number.
+	 */
+	private int getRandom(int min, int max) {
+		return (rand.nextInt((max - min)) + min);
+	}
+
+	/**
+	 * Method to load all rooms from file and return the sizes.
+	 *
+	 * <p>The URISyntaxException occurs if somehow the URL generated by {@link Class#getResource(String)}
+	 * is invalid. Since this does not occur, this exception can safely be ignored.
+	 *
+	 * @return
+	 *          An ArrayList of sizes.
+	 */
+	@SneakyThrows(URISyntaxException.class)
+	public static List<Size> loadRooms() {
+		File[] files = new File(RandomLevelFactory.class.getResource("/rooms").toURI()).listFiles();
+
+		/**
+		 * If the folder is not a folder, we throw an IllegalStateException exception.
+		 */
+		if (files == null) {
+			throw new IllegalStateException("There are no rooms.");
+		}
+
+		List<Size> sizes = new ArrayList<>();
+		for (File f : files) {
+			Matcher m = PATTERN.matcher(f.getName());
+			if (m.matches()) {
+				int width = Integer.parseInt(m.group("width"));
+				int height = Integer.parseInt(m.group("height"));
+				sizes.add(new Size(width, height));
+			}
+		}
+		return sizes;
+	}
+
+	/**
+	 * Method to create a list of {@link GeneratorRoom}.
+	 *
+	 * @return
+	 * 			the created generatorrooms
+     */
+	private List<GeneratorRoom> create() {
+		List<GeneratorRoom> rooms = new ArrayList<>();
+		List<Size> sizes = loadRooms();
+		if (!allowDuplicates && sizes.size() < amount) {
+			throw new IllegalArgumentException("You are requesting more rooms than there are.");
+		}
+
+		/**
+		 * Try to place the requested amount of rooms.
+		 */
+		for (int i = 0; i < amount; i++) {
+			int attempts = 0;
+			boolean success = false;
+			Size rSize = getRandomSizes(sizes, allowDuplicates);
+			GeneratorRoom newRoom = null;
+			if (rSize.getWidth() >= MAX_WIDTH || rSize.getHeight() >= MAX_HEIGHT) {
+				throw new IllegalArgumentException("It is impossible for this level to fit in your map size.");
+			}
+
+			/**
+			 * Try MAX_ATTEMPTS times to create the room, if all of these
+			 * fail restart the entire algorithm.
+			 */
+			while (!success && attempts < MAX_ATTEMPTS) {
+				success = true;
+				int xCoord = getRandom(0, MAX_WIDTH - rSize.getWidth());
+				int yCoord = getRandom(0, MAX_HEIGHT - rSize.getHeight());
+				newRoom = new GeneratorRoom(xCoord, yCoord, rSize);
+				for (GeneratorRoom r : rooms) {
+					if (newRoom.intersects(r)) {
+						success = false;
+						attempts++;
+					}
+				}
+			}
+
+			/**
+			 * If the room has been places, add it to the list of places rooms.
+			 * Otherwise, break out of the while loop as there is no need to continue.
+			 */
+			if (success) {
+				rooms.add(newRoom);
+			} else {
+				break;
+			}
+		}
+
+		/**
+		 * If all rooms have been placed, return the list of all the rooms.
+		 * Else call the algorithm again and start all over.
+		 */
+		if (rooms.size() == amount) {
+			return rooms;
+		} else {
+			return create();
+		}
+	}
+
+	/**
+	 * Create a matrix of TileTypes to represent the map which will be generated.
+	 *
+	 * @param rooms
+	 * 			the generated rooms
+	 *
+	 * @return
+	 * 			the generated matrix of TileTypes
+     */
+	private TileType[][] carveRooms(List<GeneratorRoom> rooms) {
+		TileType[][] carved = new TileType[MAX_WIDTH][MAX_HEIGHT];
+		for (GeneratorRoom r : rooms) {
+			for (int x = r.getxLeft(); x < r.getxRight(); x++) {
+				for (int y = r.getyLeft(); y < r.getyRight(); y++) {
+					if (x == r.getxLeft() || x == r.getxRight() - 1 || y == r.getyLeft() || y == r.getyRight() - 1) {
+						carved[x][y] = TileType.WALL;
+					} else {
+						carved[x][y] = TileType.FLOOR;
+					}
+				}
+			}
+		}
+		return carved;
+	}
+
+	private TileType[][] carveCorridors(TileType[][] map, List<GeneratorRoom> rooms) {
+		for (int i = 1; i < rooms.size(); i++) {
+			Vector2f prevCenter = rooms.get(i - 1).getCenter();
+			Vector2f currCenter = rooms.get(i).getCenter();
+			int rn = getRandom(0, 2);
+			if (rn == 1) {
+				map = hCorridor(map, prevCenter.getX(), currCenter.getX(), prevCenter.getY());
+				map = vCorridor(map, prevCenter.getY(), currCenter.getY(), currCenter.getX());
+			} else {
+				map = vCorridor(map, prevCenter.getY(), currCenter.getY(), prevCenter.getX());
+				map = hCorridor(map, prevCenter.getX(), currCenter.getX(), currCenter.getY());
+			}
+		}
+		return map;
+	}
+
+	private Size getRandomSizes(List<Size> sizes, boolean allowDuplicates) {
+		Size selected = sizes.get(getRandom(0, sizes.size()));
+		if (!allowDuplicates) {
+			sizes.remove(selected);
+		}
+		return selected;
+	}
+
+	/**
+	 * Method for generating a horizontal corridor.
+	 *
+	 * @param map
+	 * 			the map in which to generate the corridor
+	 * @param x1
+	 * 			the x coordinate of a room
+	 * @param x2
+	 * 			the x coordinate of a room
+	 * @param yF
+	 * 			the y coordinate of a room
+	 * @return
+	 * 			the map with the corridor cut out
+	 */
+	private TileType[][] hCorridor(TileType[][] map, float x1, float x2, float yF) {
+		int min = (int) Math.floor(Math.min(x1, x2));
+		int max = (int) Math.floor(Math.max(x1, x2));
+		int y = (int) Math.floor(yF);
+		for (int x = min; x < max + 1; x++) {
+			if (map[x][y] == null || map[x][y] == TileType.WALL) {
+				map[x][y] = TileType.CORRIDOR;
+			}
+		}
+		return map;
+	}
+
+	/**
+	 * Method for generating a vertical corridor.
+	 *
+	 * @param map
+	 * 			the map in which to generate the corridor
+	 * @param y1
+	 * 			the y coordinate of a room
+	 * @param y2
+	 * 			the y coordinate of a room
+	 * @param xF
+	 * 			the x coordinate of a room
+	 * @return
+	 * 			the map with the corridor cut out
+	 */
+	private TileType[][] vCorridor(TileType[][] map, float y1, float y2, float xF) {
+		int min = (int) Math.floor(Math.min(y1, y2));
+		int max = (int) Math.floor(Math.max(y1, y2));
+		int x = (int) Math.floor(xF);
+		for (int y = min; y < max + 1; y++) {
+			if (map[x][y] == null || map[x][y] == TileType.WALL) {
+				map[x][y] = TileType.CORRIDOR;
+			}
+		}
+		return map;
 	}
 
 }
