+++ src/main/java/nl/tudelft/contextproject/model/level/TileType.java
@@ -1,0 +1,10 @@
+package nl.tudelft.contextproject.model.level;
+
+/**
+ * Enum representing the possible TileTypes in the game.
+ */
+public enum TileType {
+    FLOOR,
+    WALL,
+    CORRIDOR;
+}
